support other PAF inputs
