+++ src/common/utils.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <cmath>
+#include <string>
+
+namespace wfmash {
+    bool is_a_number(const std::string& s);
+
+    int64_t handy_parameter(const std::string& value);
+}
